requirements.txt
streamlit~=1.44.1
langchain~=0.3.23
langchain-community
langchain-openai~=0.3.12
langchain-pinecone
openai
pinecone~=6.0.2
PyMuPDF
python-docx~=1.1.2
tiktoken
python-dotenv~=1.1.0
requests~=2.32.3
pinecone-client~=2.2.4
langchain-core~=0.3.51